refactor(notebook-cloud): extract sharing controls
